requirements.txt
librosa
numpy
matplotlib
python-telegram-bot
python-dotenv
pydub
assemblyai
pyTelegramBotAPI
ffmpeg-python
